pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["piltext"]

[project]
name = "piltext"
authors = [
  { name="Holger Nahrstaedt", email="nahrstaedt@gmail.com" },
]
description = "Creates PNG from text using Pillow"
keywords = ["Pillow", "png", "PIL"]
urls = { Homepage = "https://github.com/holgern/piltext" }
readme = "README.md"
requires-python = ">3.7.0"
dependencies = [
    "Pillow",
]
license = {file = "LICENSE"}
classifiers=[
    'Intended Audience :: Developers',
    'License :: OSI Approved :: MIT License',
    'Operating System :: OS Independent',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.7',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Topic :: Software Development :: Libraries',
]

dynamic=["version"]

[tool.setuptools_scm]
write_to = "piltext/_version.py"

[tool.tox]
legacy_tox_ini = """

[testenv]
deps =
    pytest
    pytest-cov[all]
commands =
    pytest
"""
